Refactor Drive auth helpers into auth module
